Polymorphic Memory Resource List added

+++ ExecutionLog.cpp
@@ -1,6 +1,6 @@
 #include "ExecutionLog.h"
 
-void ExecutionLog::insertExecutedTrades(uint64_t buyerID, uint64_t sellerID, uint64_t price, uint32_t quantity) {
+void ExecutionLog::insertExecutedTrades(uint64_t buyerID, uint64_t sellerID, uint64_t price, uint64_t quantity) {
 
 	Trade newTrade{ buyerID, sellerID, price, quantity };
