+++ ExecutionLog.h
@@ -9,6 +9,6 @@ private:
 	std::vector<Trade> TradesExecuted;
 
 public:
-	void insertExecutedTrades(uint64_t buyerID, uint64_t sellerID, uint64_t price, uint32_t quantity);
+	void insertExecutedTrades(uint64_t buyerID, uint64_t sellerID, uint64_t price, uint64_t quantity);
 	void printLog();
 };
